add PatchContext mechanism for efficiently applying a series of updates
This commit modifies the DBMS API to indroduce the concept of a patch
context, which may be used to incrementally build a new database
revision by applying a series of inserts, updates, and/or deletes to a
base revision.  This mechanism allows the implementation to preserve
state across multiple updates and optimize execution of related
updates.

This commit also removes the Query and Patch types from the API,
eliminating the extra steps necessary to execute a query and apply a
patch, respectively.