fix reliance on undefined iteration order in SQLServer.makeOptionalColumnList
Iterating over a HashMap will visit the entries in undefined order, so
we need to remember the order in which a table's columns are declared
by storing them in a List as well as a Map.  This allows us to make
sense of inserts which do not specify an explicit column list.