allow columns, indexes, and foreign keys to be added dynamically
Previously, the set of columns and indexes for a given table was
immutable, and there was no natural way to link tables together via
foreign keys without making tables at least temporarily mutable.  This
commit modifies the DBMS API and implementation to such items to be
added and removed when defining a new revision, just like the data
itself.

There is no longer a fixed set of columns, indexes, or foreign keys
associated with a table, except that the table must have an immutable
primary key and each row must have values for at least the columns
listed in the primary key.  When a row is visited as part of a query
result, any column for which the row has no value is considered null.