change signature of DBMS.ConflictResolver.resolveConflict
This commit changes the signature of resolveConflict so it accepts an
array containing primary key values which uniquely identify the
conflicting row.  It also removes the base, left, and right Revision
parameters since those are provided by the caller of DBMS.merge and
are therefore already available in the context in which the merge
takes place.