lots of bugfixes related to the epidemic replication server
The biggest change is to allow naming of columns and tables so that
the the replication servers on remote nodes can match up incoming
diffs to local tables and columns.

This commit also removes the distinction between IndexKeys and Indexes
in MyDBMS, since it added complication and didn't really make the DBMS
interface any more intuitive.

Finally, there is a new query method in DBMS for lightweight queries
of single columns based on a unique index key.