fix operations involving numeric literals and 32-bit integer columns
When parsing queries in SQLServer, we parse numeric literals as 64-bit
values, since that's the largest integer type the server supports.
However, this led to ClassCastExceptions when we tried, e.g., to
insert values of type Long into a column of type Integer.

This commit introduces the Coerceable interface, allowing values to be
implicitly converted to other types as necessary.  We use this
interface to allow integer values of different types to be converted
between each other.