optionally skip deleted foreign key references when computing diffs
For many applications, there is no reason for a database-wide diff to
explicitly visit deleted rows with one or more broken foreign key
references to other tables, since their deletion is implied by the
deletion or update of the respective referent row(s).  This is
especially important when serializing diffs with one-to-many
relationships, since it allows us to reduce the size of the diff by
omitting such redundant deletes.

This commit also deprecates the DBMS and MyDBMS, since the
functionality they provide has been moved elsewhere.  We'll want to
remove them entirely before the first stable release, but that will
require updating the test suites that use them.