fix spurious conflicts in EpidemicServer due to unacknowledged merges
In addition to acknowledging incoming diffs, we must also "ack the
acks", i.e. acknowledge that another node has merged a diff so we can
properly build on that merge later on as further diffs and acks
arrive.  Otherwise, we get spurious conflicts when calculating merges
where the base revision used is actually an ancestor of the revision
we ought to use.

The tricky part here is that we need to short-circuit this process so
that we don't generate an infinite cascade of acks, each one
triggering another.  Hence the logic in EpidemicServer.acceptAck which
only "acks an ack" if the ack being acked is not itself an ack of an
ack.  Simple, right?