implement new approach to coercing literals in SQLServer
The problem with the previous strategy based on the Coerceable
interface was that it required the schema to use types which implement
that interface instead of the more familiar primitive types like
Integer and Long.  This is an onerous requirement for the case of an
exisiting schema to which one would like to add an SQL interface.

The new approach eliminates the Coerceable interface and moves the
coercion to SQLServer itself, where we can accomodate types like
Integer and Long.  Custom literals require the use of the Stringable
interface, which is a reasonable requirement for those who want to
input data using custom column types from a command-line interface.