add Column.comparator to allow custom ordering of column values
We can use this to implement descending ordering in queries with
order-by clauses, as well as allow views of tables with different
ordering than what is defined by the compareTo methods of their
respective columns (e.g. sorting Strings using compareToIgnoreCase).
It also removes the restriction that column types must implement
Comparable as long as a suitable Comparator is explicitly supplied.