add ContextRowListener for use with DiffMachine
This can be useful when some kind of state needs to be passed through
from the code asking for the next diff event (i.e. the caller of
DiffMachine.next) to the RowListener which handles the event.  One
example is when the event must be written to a stream, in which case
the object passed woult be the stream itself.